[FIX] mail: restrict invite input autofocus
Before of this commit:

- The invite input always tried to autofocus, stealing focus from dialogs like
the call-permission dialog.

After of this commit:

- Autofocus only when the input exists and the context is active.
- Moved mockPermissionsPrompt to mail_test_helpers.

Part of task-5227387

closes odoo/odoo#236248
