+++ addons/mail/static/tests/discuss/core/channel_invitation.test.js
@@ -3,12 +3,14 @@ import {
     contains,
     defineMailModels,
     insertText,
+    mockPermissionsPrompt,
     openDiscuss,
     setupChatHub,
     start,
     startServer,
 } from "@mail/../tests/mail_test_helpers";
 import { describe, test } from "@odoo/hoot";
+import { animationFrame } from "@odoo/hoot-dom";
 import { mockDate } from "@odoo/hoot-mock";
 import { Command, getService, serverState, withUser } from "@web/../tests/web_test_helpers";
 
@@ -244,4 +246,15 @@ test("Invite sidebar action has the correct title for group chats", async () => 
     await click("button[title='Chat Actions']");
     await click(".o-dropdown-item", { text: "Invite People" });
     await contains(".modal-title", { text: "Mitchell Admin and Demo" });
+});
+
+test("Active dialog retains focus over invite input", async () => {
+    await startServer();
+    mockPermissionsPrompt();
+    await start();
+    await openDiscuss();
+    await click("button[title='New Meeting']");
+    await animationFrame();
+    await contains(".o-discuss-ChannelInvitation");
+    await contains("button:focus", { text: "Use Camera" });
 });
